requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.3
coverage==7.10.3
cycler==0.12.1
decorator==5.2.1
executing==2.2.0
Flask==3.1.0
fonttools==4.59.2
frozenlist==1.7.0
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jsonpickle==4.1.1
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
memory-profiler==0.61.0
multidict==6.6.4
multipledispatch==1.0.0
natsort==8.4.0
networkx==3.5
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pandas_flavor==0.7.0
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
playwright==1.49.1
pluggy==1.5.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==6.1.1
psycopg==3.2.4
ptyprocess==0.7.0
pure_eval==0.2.3
pyee==12.0.0
Pygments==2.19.2
pyjanitor==0.31.0
PyJWT==2.10.1
pyparsing==3.2.4
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.6.2
pytest-xprocess==1.0.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
pyvis==0.3.2
requests==2.32.3
scipy==1.16.1
six==1.17.0
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
twilio==9.7.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
xarray==2025.8.0
yarl==1.20.1
